Refactor test Stubs to Workbench jobs and PHPUnit Config
This commit refactors the namespace and location of test stubs and moves them from 'ViicSlen\TrackableTasks\Tests\Stub' to 'Workbench\App\Jobs.' This move is due to the realignment of project structure for better organization of the code. Single-run queue work commands in tests have been replaced with the '--stop-when-empty' flag for complete processing of all jobs. Additionally, significant changes were made to 'phpunit.xml.dist'.